requirements.txt
python-telegram-bot
websocket-client
undetected_chromedriver
pyvirtualdisplay